+++ .bud/Build.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using Bud;
 using Bud.CSharp;
 using Bud.Projects;
@@ -30,10 +31,16 @@ public class Build : IBuild {
         Cs.Dependency("NUnit")
       ))
       .Project("Bud.SystemTests", Path.Combine(baseDir, "SystemTests"), Cs.Test(
-        Cs.Dependency("Bud.Test")
+        Cs.Dependency("Bud.Test"),
+        NUnitTestTargetPlugin.NUnitArgumentsKey.Modify(RunTestsWithouShadowDlls)
       ))
       .Project("Bud.Examples.Snippets", Path.Combine(baseDir, "Bud.Examples.Snippets"), Cs.Dll(
         Cs.Dependency("Bud.Core")
       ));
+  }
+
+  private static List<string> RunTestsWithouShadowDlls(List<string> oldArgs) {
+    oldArgs.Add("/noshadow");
+    return oldArgs;
   }
 }