+++ Test/src/main/cs/Util/TestProjects.cs
@@ -1,7 +1,11 @@
+using System;
 using System.IO;
 
 namespace Bud.Test.Util {
   public static class TestProjects {
+    public static readonly string TestAssemblyDir = Path.GetDirectoryName(typeof(TestProjects).Assembly.Location);
+    public static readonly string TestProjectsBaseDir = Path.Combine(TestAssemblyDir, "..", "..", "..", "..", "..", "src", "test", "budProjects");
+
     public static TemporaryDirBuildCommander LoadBuildCommander(string projectDirName) {
       return new TemporaryDirBuildCommander(TemporaryCopy(projectDirName));
     }
@@ -19,7 +23,7 @@ namespace Bud.Test.Util {
     }
 
     public static string GetPathOfProject(string projectDirName) {
-      return Path.Combine(Directory.GetCurrentDirectory(), "..", "..", "..", "..", "..", "src", "test", "budProjects", projectDirName);
+      return Path.Combine(TestProjectsBaseDir, projectDirName);
     }
 
     public static TemporaryDirBuildCommander LoadBuildCommander(object systemTestClassInstance) {